CodeCleanup. Changed names to comply with required scheme Dates in Dto changed from Integer to Long, to avoid unnecessary casting. Typos corrected. TODO: docstrings

+++ course_project/src/main/java/yetAnotherWeatherSource/api/dto/ListDto.java
@@ -6,6 +6,6 @@ import org.codehaus.jackson.annotate.JsonIgnoreProperties;
 @Data
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class ListDto {
-    private Integer dt;
+    private Long dt;
     private MainDto main;
 }
